reload login

+++ ubuntu/install.sh
@@ -1,7 +1,7 @@
 #!/bin/bash -e
 #Docker Installer
 #author: elmerfdz
-version=v0.41.6-3
+version=v0.41.6-4
 
 #Script Requirements
 prereqname=('Curl' )
@@ -233,13 +233,28 @@ additional_docker_config()
             echo
             echo 'dc="docker-compose -f '"/opt/docker/docker-compose.yml"'"' >> /etc/environment
             echo "Done!"
+            echo
+            echo "You may have to log out and log back in for the changes to take effect."
         else
             echo "Skipped"
         fi
-        echo -e "\e[1;36m> \e[0mPress any key to return to menu..."
+        
+        if [ $dc_dcom_var = "Y" ] || [ $dc_dcom_var = "y" ];
+        then
         rm -rf ./inst_4_temp
+        echo -e "\e[1;36m> \e[0mPress any key to quit the script and refresh login"
         read
-        shell_reload           
+        su - $SUDO_USER
+
+        elif [ $dc_dcom_var = "N" ] || [ $dc_dcom_var = "n" ];
+        then
+        rm -rf ./inst_4_temp
+        echo -e "\e[1;36m> \e[0mPress any key to continue..."
+        read
+        shell_reload   
+        fi
+
+        
 	}        
 
 # Debug env vars
